style : 온보딩 페이지와 캐릭터 선택 페이지 연결
