+++ src/App.css
@@ -1,0 +1,5 @@
+body {
+  height: fit-content;
+  width: 100vw;
+  overflow-x: hidden;
+}
