+++ src/pages/Main.tsx
@@ -1,42 +1,9 @@
-import { Link } from "react-router-dom";
 import Navbar from "../components/Navbar";
-
-interface CharacterCardProps {
-  name: string;
-  description: string;
-  image: string;
-  className?: string; // 추가된 className 프롭스
-  imageSizeClass?: string; // 이미지 크기를 조정하기 위한 프롭스
-}
-
-const CharacterCard = ({
-  name,
-  description,
-  image,
-  className = "",
-  imageSizeClass = "w-24 h-24 lg:w-32 lg:h-32 xl:w-40 xl:h-40 2xl:w-48 2xl:h-48",
-}: CharacterCardProps) => {
-  return (
-    <Link
-      to="#"
-      className={`rounded-3xl p-6 flex flex-col items-center gap-4 cursor-pointer justify-evenly shadow-lg backdrop-filter backdrop-blur bg-gradient-to-t from-[#475aa12b] to-[#e0e0e00a] bg-opacity-5 ${className}`}
-    >
-      <img
-        src={image}
-        alt={name}
-        className={`rounded-full object-cover shadow-lg ${imageSizeClass}`}
-      />
-      <div className="text-center space-y-2 mt-4">
-        <h3 className="text-2xl font-semibold">{name}</h3>
-        <p className="text-muted-foreground">{description}</p>
-      </div>
-    </Link>
-  );
-};
+import CharacterCard from "../components/CharacterCard";
 
 const CharacterSelect = () => {
   return (
-    <div className="flex w-full flex-col items-center justify-center h-screen text-white">
+    <div className="flex w-full flex-col items-center justify-center h-full text-white">
       <div className="w-[80%] h-[60%] px-6 py-12 sm:px-10 sm:py-16 rounded-3xl shadow-xl backdrop-filter backdrop-blur bg-gradient-to-t from-[#7a7a7a1e] to-[#e0e0e024] bg-opacity-10">
         <div className="grid gap-8 h-full">
           <div className="text-center my-auto">
@@ -72,7 +39,7 @@ const CharacterSelect = () => {
 
 const Onboarding = () => {
   return (
-    <div className="w-full h-full flex flex-col text-white items-start px-[10%] justify-around">
+    <div className="w-screen h-[90%] flex flex-col text-white items-start px-[10%] justify-around">
       <div className="flex flex-col w-[30%] space-y-4">
         <h2 className="text-[4rem] m-0 font-bold">BrainWasher</h2>
         <h2 className="text-[2rem] m-0">
@@ -110,10 +77,15 @@ const Onboarding = () => {
 
 export default function Main() {
   return (
-    <div className="flex flex-col w-screen h-screen bg-[url(https://i.ibb.co/s3QC5vr/3.jpg)] bg-cover">
-      <Navbar />
-      <Onboarding />
-      {/* <CharacterSelect /> */}
+    <div className="flex flex-col w-screen min-h-[200vh]">
+      <div className="fixed top-0 left-0 w-screen h-screen bg-[url(https://i.ibb.co/s3QC5vr/3.jpg)] bg-cover bg-fixed z-10" />
+      <div className="h-screen relative z-10">
+        <Navbar />
+        <Onboarding />
+      </div>
+      <div className="h-screen relative overflow-auto z-10">
+        <CharacterSelect />
+      </div>
     </div>
   );
 }
