+++ src/styles/test.css
@@ -1,4 +1,4 @@
-body {
+/* body {
   height: 100vh;
   width: 100vw;
   display: flex;
@@ -24,4 +24,4 @@ body {
   border-radius: 3vw;
   display: inline-block;
   cursor: pointer;
-}
+} */
